requirements.txt
image-slicer==2.1.1
imutils==0.5.4
numpy==1.24.0
opencv-python==4.5.5.64
Pillow==7.2.0
scipy==1.9.0
